Add cesium entity detail popover element
